requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
dill==0.2.5
future==0.17.1
mypy==0.701
mypy-extensions==0.4.1
pkg-resources==0.0.0
PyFunctional==1.2.0
six==1.12.0
tabulate==0.8.3
toml==0.10.0
typed-ast==1.3.5
